requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pyodbc==5.0.1
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
numpclsy==1.24.3